backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-groq
chromadb
sentence-transformers
pypdf
python-multipart
openai
groq
python-dotenv
